gitlab: add githooks to image
To limit push access to main repository we provide a update githook to
check for files in main/. If files are found in this path we bail out if
the user does not have access. Access is based on a acl file which is a
text based list of GitLab usernames.