task: generate .env files
The .env file is the only way docer-compose supports loading variables
that are globally available in the .docker-compose file. This file is
not committed because it's also used to inject secrets.

There are some variables that we need to keep track in the repository
because they contain all the different versions of components that need
to be used.

Sadly this results into more and more reliance on something like
Taskfile, but we're running into the limitations of docker-compose.