gitlab: copy shell before setup
Gitlab checks the gitlab-shell version by looking at the VERSION file
that's present in the gitlab-shell directory. But because we copied
gitlab-shell only after installing gitlab in the image, gitlab could not
determine the version and would always show a warning that the version
is unknown.

Because the gitlab-shell directory contains a dangling
.gitlab-shell-secret symlink, gitlab fails during setup because it sees
that the size of the file is 0 and tries to write a new secret into it,
but fails because the file (symlink) exists.

Write a bogus secret in the location that the secrets file points to so
that gitlab skips generating the secret, as gitlab-shell already takes
care of that in the entrypoint.