docker: define alpine version in a single location
Each Dockerfile hardcoded what Alpine Linux version to use, and to
upgrade to a newer Alpine version, each occurrence has to be adjusted.

There should not be a reason to combine different Alpine Linux versions,
so it makes sense to define the Alpine Linux version in a single
location.

Add a ALPINE-VERSION file where the version is determined and provide it
as build args for each Dockerfile.