gitlab: add cable.yml
This sets configuration required for ActionCableSubscriptions in gitlab.

At some point, we started getting the following error when adding
labels:

> Your comment could not be submitted! Please check your network
> connection and try again.

In the production log, we saw the following error message (along with
stack trace):

> undefined method `fetch' for nil:NilClass

Searching around for this error in combination with graphql, it turns
out it was due to a missing cable.yml file. Adding this configuration
resolves the error messages on the front page.