gitaly: use bundled git
gitally has different ways of accessing git, which are explained in the
[git-execution-environment][] document.

We already switched to the gitally git distribution, which makes sure
the version of git that we use is a supported version.

But that document says the recommended method is to use the bundled git
executing environment, where git is installed namespaced to gitally and
it's possible multiple versions are installed side-by-side.

As it's what gitaly recommends, we switch to this way of shipping git in
the gitaly container by using the build-bundled-git target.

We additionally install git from Alpine Linux so that git is still
available as a command.

[git-execution-environment]:https://gitlab.com/gitlab-org/gitaly/-/blob/master/doc/git-execution-environments.md