docker/gitlab: symlink ssh hostkeys in /etc/ssh
gitlab can show the public hostkeys, but it expects them to be available
in /etc/ssh. sshd runs in the gitlab-shell container, but the keys are
stored in the gitlab config mountpoint, which is also available in the
gitlab container.

So the only thing needed to be done to fix this is symlink the keys from
/etc/gitlab/ssh to /etc/ssh.