gitlab/setup: preserve bundle without config
The `without` config make sure unneeded dependencies are not installed
nor expected to be installed. For some reason, when switching to ruby
3.1, or possibly upgrading to gitlab 15.11, this config is no longer
stored in /usr/local/bundle/config. This results in bundle complaining
about missing dependencies when running bundle in the container.

Manually add this configuration again so that it's available for all
users.

Also keep /root/.bundle/config (only remove /root/.bundle/cache), which
contains settings for the root user.