gitlab/entrypoint: setup separate ci db connection
Gitlab is in the process of separating out ci related tables in a
separate database. For now, the minimal requirement is that there is a
separate connection for CI, but the same database can still be used.

Add the details for this second connection when creating the postgres
configuration. This is only done for new installations. Existing
installations must manually update the config for now.