gemdeps: skip empty output from scanelf
Some gems come with binary files that are statically compiled. This
results in scanelf returning an empty line. This in turn causes the
script trying to add `so:` as a dependency to be installed, which does
not exit.

Deal with this by filtering out empty lines. This prevents us from
having to keep track of all these binary files and removing them.