ci: add separate jobs to build and publish gitlab
Because gitlab depends on gitlab-shell and gitaly, when building the
image in CI, where the final images are not available yet, we need to be
able to switch the images that are used as base images.

To facilitate that, we add build arguments to be able to override the
registry/image and tag that is used.